# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorMessage {
  error: string;
}

/**
 * Wall construction system
 */
export type ProjectWallSystem = typeof ProjectWallSystem[keyof typeof ProjectWallSystem];


export const ProjectWallSystem = {
  bahareque: 'bahareque',
  tapia_pisada: 'tapia_pisada',
  adobe: 'adobe',
  guadua_vista: 'guadua_vista',
} as const;

export type ProjectRoofType = typeof ProjectRoofType[keyof typeof ProjectRoofType];


export const ProjectRoofType = {
  teja_barro: 'teja_barro',
  palma: 'palma',
  zinc: 'zinc',
  techo_verde: 'techo_verde',
} as const;

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  diseno: 'diseno',
  cotizado: 'cotizado',
  construccion: 'construccion',
  terminado: 'terminado',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  /** Wall construction system */
  wallSystem: ProjectWallSystem;
  roofType: ProjectRoofType;
  status: ProjectStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type ProjectInputWallSystem = typeof ProjectInputWallSystem[keyof typeof ProjectInputWallSystem];


export const ProjectInputWallSystem = {
  bahareque: 'bahareque',
  tapia_pisada: 'tapia_pisada',
  adobe: 'adobe',
  guadua_vista: 'guadua_vista',
} as const;

export type ProjectInputRoofType = typeof ProjectInputRoofType[keyof typeof ProjectInputRoofType];


export const ProjectInputRoofType = {
  teja_barro: 'teja_barro',
  palma: 'palma',
  zinc: 'zinc',
  techo_verde: 'techo_verde',
} as const;

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  diseno: 'diseno',
  cotizado: 'cotizado',
  construccion: 'construccion',
  terminado: 'terminado',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  location?: string;
  wallSystem: ProjectInputWallSystem;
  roofType: ProjectInputRoofType;
  status?: ProjectInputStatus;
  notes?: string;
}

export type ProjectUpdateWallSystem = typeof ProjectUpdateWallSystem[keyof typeof ProjectUpdateWallSystem];


export const ProjectUpdateWallSystem = {
  bahareque: 'bahareque',
  tapia_pisada: 'tapia_pisada',
  adobe: 'adobe',
  guadua_vista: 'guadua_vista',
} as const;

export type ProjectUpdateRoofType = typeof ProjectUpdateRoofType[keyof typeof ProjectUpdateRoofType];


export const ProjectUpdateRoofType = {
  teja_barro: 'teja_barro',
  palma: 'palma',
  zinc: 'zinc',
  techo_verde: 'techo_verde',
} as const;

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  diseno: 'diseno',
  cotizado: 'cotizado',
  construccion: 'construccion',
  terminado: 'terminado',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  wallSystem?: ProjectUpdateWallSystem;
  roofType?: ProjectUpdateRoofType;
  status?: ProjectUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export type RoomKind = typeof RoomKind[keyof typeof RoomKind];


export const RoomKind = {
  habitacion: 'habitacion',
  sala: 'sala',
  cocina: 'cocina',
  bano: 'bano',
  comedor: 'comedor',
  corredor: 'corredor',
  taller: 'taller',
  otro: 'otro',
} as const;

export interface Room {
  id: number;
  projectId: number;
  name: string;
  kind: RoomKind;
  /** Width in meters */
  widthM: number;
  /** Length in meters */
  lengthM: number;
  /** Wall height in meters */
  heightM: number;
  /**
     * X position in meters on the floor plan
     * @nullable
     */
  posX?: number | null;
  /**
     * Y position in meters on the floor plan
     * @nullable
     */
  posY?: number | null;
}

export type RoomInputKind = typeof RoomInputKind[keyof typeof RoomInputKind];


export const RoomInputKind = {
  habitacion: 'habitacion',
  sala: 'sala',
  cocina: 'cocina',
  bano: 'bano',
  comedor: 'comedor',
  corredor: 'corredor',
  taller: 'taller',
  otro: 'otro',
} as const;

export interface RoomInput {
  /** @minLength 1 */
  name: string;
  kind: RoomInputKind;
  /** @exclusiveMinimum 0 */
  widthM: number;
  /** @exclusiveMinimum 0 */
  lengthM: number;
  /** @exclusiveMinimum 0 */
  heightM: number;
  posX?: number;
  posY?: number;
}

export type RoomUpdateKind = typeof RoomUpdateKind[keyof typeof RoomUpdateKind];


export const RoomUpdateKind = {
  habitacion: 'habitacion',
  sala: 'sala',
  cocina: 'cocina',
  bano: 'bano',
  comedor: 'comedor',
  corredor: 'corredor',
  taller: 'taller',
  otro: 'otro',
} as const;

export interface RoomUpdate {
  /** @minLength 1 */
  name?: string;
  kind?: RoomUpdateKind;
  /** @exclusiveMinimum 0 */
  widthM?: number;
  /** @exclusiveMinimum 0 */
  lengthM?: number;
  /** @exclusiveMinimum 0 */
  heightM?: number;
  /** @nullable */
  posX?: number | null;
  /** @nullable */
  posY?: number | null;
}

export type MaterialCategory = typeof MaterialCategory[keyof typeof MaterialCategory];


export const MaterialCategory = {
  estructura: 'estructura',
  muros: 'muros',
  cubierta: 'cubierta',
  acabados: 'acabados',
  otros: 'otros',
} as const;

export interface Material {
  id: number;
  name: string;
  /** Unit of measure, e.g. unidad, m3, bulto, m2 */
  unit: string;
  /** Price per unit in COP */
  unitPrice: number;
  category: MaterialCategory;
}

export type MaterialInputCategory = typeof MaterialInputCategory[keyof typeof MaterialInputCategory];


export const MaterialInputCategory = {
  estructura: 'estructura',
  muros: 'muros',
  cubierta: 'cubierta',
  acabados: 'acabados',
  otros: 'otros',
} as const;

export interface MaterialInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  unitPrice: number;
  category: MaterialInputCategory;
}

export type MaterialUpdateCategory = typeof MaterialUpdateCategory[keyof typeof MaterialUpdateCategory];


export const MaterialUpdateCategory = {
  estructura: 'estructura',
  muros: 'muros',
  cubierta: 'cubierta',
  acabados: 'acabados',
  otros: 'otros',
} as const;

export interface MaterialUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  unit?: string;
  /** @minimum 0 */
  unitPrice?: number;
  category?: MaterialUpdateCategory;
}

export type EstimateItemCategory = typeof EstimateItemCategory[keyof typeof EstimateItemCategory];


export const EstimateItemCategory = {
  estructura: 'estructura',
  muros: 'muros',
  cubierta: 'cubierta',
  acabados: 'acabados',
  otros: 'otros',
} as const;

export interface EstimateItem {
  materialName: string;
  category: EstimateItemCategory;
  quantity: number;
  unit: string;
  /** 0 when the material is not in the catalog */
  unitPrice: number;
  subtotal: number;
  /** False when no catalog price was found */
  priced: boolean;
}

export interface Estimate {
  projectId: number;
  floorAreaM2: number;
  wallAreaM2: number;
  roofAreaM2: number;
  items: EstimateItem[];
  totalCost: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardSummary {
  totalProjects: number;
  totalRooms: number;
  totalFloorAreaM2: number;
  byStatus: StatusCount[];
}