# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StatusCount } from './statusCount';

export interface DashboardSummary {
  totalProjects: number;
  totalRooms: number;
  totalFloorAreaM2: number;
  byStatus: StatusCount[];
}