# lib/api-zod/src/generated/types/estimate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EstimateItem } from './estimateItem';

export interface Estimate {
  projectId: number;
  floorAreaM2: number;
  wallAreaM2: number;
  roofAreaM2: number;
  items: EstimateItem[];
  totalCost: number;
}