# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MaterialInputCategory } from './materialInputCategory';

export interface MaterialInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  unitPrice: number;
  category: MaterialInputCategory;
}