# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomKind } from './roomKind';

export interface Room {
  id: number;
  projectId: number;
  name: string;
  kind: RoomKind;
  /** Width in meters */
  widthM: number;
  /** Length in meters */
  lengthM: number;
  /** Wall height in meters */
  heightM: number;
  /**
     * X position in meters on the floor plan
     * @nullable
     */
  posX?: number | null;
  /**
     * Y position in meters on the floor plan
     * @nullable
     */
  posY?: number | null;
}